requirements.txt
pyqrcode==1.2.1
Pillow
sphinx
sphinx-rtd-theme
wheel
twine
